Defer external data fetching to a background task and return response to client straight away with nothing
External data is only meant to support existing report data. It is not a requirement. If a user is the first to hit an endpoint, then the fetch of the external data should be deferred to a background task and we should return a response to them quickly rather than holding up game reports because we are fetching external data also.

Note: I'm also removing SDG reviews from the external data because I do not think it is adding value to the platform. This was only added in an agreement with SDG that the Deck Settings reports would be added to their app, and that was never done.