Use 2 different GH tokens for the API and backend tasks
A rate limit on the API should not prevent the background tasks from still updating the redis cache with data.

Increase default cache time. With the background tasks, we no longer need these to be as regular.

Move some additional github repo data to be fetch and updated on a schedule by the background tasks service.